requirements.txt
Django==1.7.7
Pygments==2.0.2
autopep8==1.1.1
coverage==3.7.1
execnet==1.3.0
flake8==2.4.0
flake8-import-order==0.6
flake8-quotes==0.0.1
httpie==0.9.2
jsonfield==1.0.3
mccabe==0.3
mock==1.0.1
model-mommy==1.2.4
pep8==1.5.7
pep8-naming==0.2.2
psycopg2==2.6
py==1.4.26
pyflakes==0.8.1
pytest==2.7.0
pytest-django==2.8.0
pytest-xdist==1.11
python-dateutil==2.4.2
requests==2.6.0
responses==0.3.0
six==1.9.0
